refactoring- consolidated SapphirePolicy and DefaultSapphireCallImpls
